Add integration tests
This adds integration tests to ppx_blob. The choice of test cases is the
following:

- Dune 1.x with `(lang dune 1.0)`
- Dune 3.x with `(lang dune 3.0)`
- OCamlbuild
- Direct use of `ocamlopt`

Each test directory is self-contained and has its own Opam dependencies.
This commit adds the corresponding GitHub Actions jobs.

The result of the tests is the following:

As already noted in other discussions, there are cases that worked with
`(lang dune 1.0)` which stopped working in `(lang dune 3.0)`. More
specifically, finding a file relative to the source file no longer works
in 3.0. This also means that if the same relative path works for a file
relative to the build directory, it will be picked instead, thus
violating the precedence rule (see readme of ppx_blob).

The tests highlight other surprising behavior:

- Absolute paths can be shadowed by relative paths (`dune-lang-1` and
  `ocamlopt`).
- ppx_blob is of impractical with OCamlbuild (you have to look for the
  source files outside the build directory).

One thing I wasn't able to test is the behavior of Merlin and OCaml-LSP.
It looks like it can be confused by the use of
working-directory-relative paths.